+++ service/history/queues/queue.go
@@ -50,13 +50,6 @@ type (
 		// as that will lead to a cycle dependency issue between shard and workflow package.
 		CreateProcessor(shard shard.Context, engine shard.Engine, cache workflow.Cache) Processor
 	}
-
-	ReplicationExecutorFactory interface {
-		CreateExecutor(shard shard.Context, engine shard.Engine, cache workflow.Cache) ReplicationExecutor
-	}
-
-	ReplicationExecutor interface {
-	}
 )
 
 const (
